NO-ISSUE: split kube helpers file
Since it is a large file that will only get bigger,
it will be better to split it to a sub package.
